+++ ci/release_notes.md
@@ -1,2 +1,5 @@
 ## Features
 - add hloc query to graphql
+
+## Improvements
+- better responsiveness of api during bootstrapping via bulk insert of trade history